requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --no-emit-options
#
alembic==1.7.7
    # via -r requirements.in
amqp==5.1.0
    # via kombu
anyio==3.5.0
    # via starlette
asgiref==3.5.0
    # via uvicorn
async-timeout==4.0.2
    # via redis
asyncpg==0.25.0
    # via -r requirements.in
attrs==21.4.0
    # via
    #   hypothesis
    #   hypothesis-graphql
    #   jsonschema
    #   pytest
    #   schemathesis
billiard==3.6.4.0
    # via celery
celery==5.2.6
    # via
    #   -r requirements.in
    #   pytest-celery
certifi==2021.10.8
    # via requests
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.0.12
    # via requests
click==8.1.2
    # via
    #   -r requirements.in
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
    #   rich-click
    #   schemathesis
    #   uvicorn
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
colorama==0.4.4
    # via schemathesis
commonmark==0.9.1
    # via rich
coverage[toml]==6.3.2
    # via pytest-cov
curlify==2.2.1
    # via schemathesis
deprecated==1.2.13
    # via redis
distlib==0.3.4
    # via virtualenv
fastapi==0.75.1
    # via -r requirements.in
filelock==3.6.0
    # via virtualenv
flake8==4.0.1
    # via -r requirements.in
graphql-core==3.2.0
    # via hypothesis-graphql
greenlet==1.1.2
    # via sqlalchemy
gunicorn==20.1.0
    # via -r requirements.in
h11==0.13.0
    # via uvicorn
hypothesis==6.41.0
    # via
    #   hypothesis-graphql
    #   hypothesis-jsonschema
    #   schemathesis
hypothesis-graphql==0.5.1
    # via schemathesis
hypothesis-jsonschema==0.22.0
    # via schemathesis
identify==2.4.12
    # via pre-commit
idna==3.3
    # via
    #   anyio
    #   requests
    #   yarl
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via -r requirements.in
jinja2==3.1.1
    # via sqladmin
jsonschema==4.4.0
    # via
    #   hypothesis-jsonschema
    #   schemathesis
junit-xml==1.9
    # via schemathesis
kombu==5.2.4
    # via celery
mako==1.2.0
    # via alembic
markupsafe==2.1.1
    # via
    #   jinja2
    #   mako
    #   wtforms
mccabe==0.6.1
    # via flake8
mirakuru==2.4.2
    # via pytest-postgresql
multidict==6.0.2
    # via yarl
nodeenv==1.6.0
    # via pre-commit
numpy==1.22.3
    # via pandas
packaging==21.3
    # via
    #   pytest
    #   redis
pandas==1.4.2
    # via -r requirements.in
pep517==0.12.0
    # via pip-tools
pip-tools==6.6.0
    # via -r requirements.in
platformdirs==2.5.1
    # via virtualenv
pluggy==1.0.0
    # via pytest
port-for==0.6.2
    # via pytest-postgresql
pre-commit==2.18.1
    # via -r requirements.in
prompt-toolkit==3.0.29
    # via click-repl
psutil==5.9.0
    # via mirakuru
py==1.11.0
    # via pytest
pycodestyle==2.8.0
    # via flake8
pydantic==1.9.0
    # via
    #   fastapi
    #   sqlmodel
pyflakes==2.4.0
    # via flake8
pygments==2.11.2
    # via rich
pyparsing==3.0.7
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
pytest==7.1.1
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-postgresql
    #   pytest-subtests
    #   schemathesis
pytest-celery==0.0.0
    # via -r requirements.in
pytest-cov==3.0.0
    # via -r requirements.in
pytest-postgresql==3.1.3
    # via -r requirements.in
pytest-subtests==0.7.0
    # via schemathesis
python-dateutil==2.8.2
    # via pandas
python-dotenv==0.20.0
    # via -r requirements.in
python-multipart==0.0.5
    # via sqladmin
pytz==2022.1
    # via
    #   celery
    #   pandas
pyyaml==6.0
    # via
    #   pre-commit
    #   schemathesis
redis==4.2.2
    # via -r requirements.in
requests==2.27.1
    # via
    #   curlify
    #   schemathesis
rich==12.2.0
    # via rich-click
rich-click==1.3.0
    # via -r requirements.in
schemathesis==3.13.8
    # via -r requirements.in
six==1.16.0
    # via
    #   click-repl
    #   junit-xml
    #   python-dateutil
    #   python-multipart
    #   virtualenv
sniffio==1.2.0
    # via anyio
sortedcontainers==2.4.0
    # via hypothesis
sqladmin==0.1.7
    # via -r requirements.in
sqlalchemy==1.4.35
    # via
    #   -r requirements.in
    #   alembic
    #   sqladmin
    #   sqlmodel
sqlalchemy2-stubs==0.0.2a21
    # via sqlmodel
sqlmodel==0.0.6
    # via -r requirements.in
starlette==0.17.1
    # via
    #   fastapi
    #   schemathesis
    #   sqladmin
toml==0.10.2
    # via pre-commit
tomli==2.0.1
    # via
    #   coverage
    #   pep517
    #   pytest
typing-extensions==4.1.1
    # via
    #   pydantic
    #   schemathesis
    #   sqlalchemy2-stubs
urllib3==1.26.9
    # via requests
uvicorn==0.17.6
    # via -r requirements.in
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
virtualenv==20.14.0
    # via pre-commit
wcwidth==0.2.5
    # via prompt-toolkit
werkzeug==2.1.1
    # via schemathesis
wheel==0.37.1
    # via pip-tools
wrapt==1.14.0
    # via deprecated
wtforms==3.0.1
    # via sqladmin
yarl==1.7.2
    # via schemathesis

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
